Harden async cmdlet pipeline lifecycle
